pyproject.toml
[tool.poetry]
name = "songmap"
version = "0.1.0"
description = ""
authors = ["ncc <ncc@github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
